requirements.txt
# Core Deep Learning & ML
tensorflow>=2.12.0
numpy>=1.23.0
pillow>=9.5.0

# Backend API
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0

# Data Processing & Visualization
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Optional: For production deployment
gunicorn>=21.0.0

